% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonify.R
\name{sonify.histogram}
\alias{sonify.histogram}
\title{Sonification of histogram objects}
\usage{
\method{sonify}{histogram}(x, ...)
}
\arguments{
\item{x}{An object of class histogram}

\item{...}{additional arguments passed to sonify}
}
\value{
The synthesized sound is returned as a `tuneR::WaveMC` object.
}
\description{
S3 method to sonify histogram objects
}
