% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonify.R
\name{sonify}
\alias{sonify}
\title{Sonify data}
\usage{
sonify(x = NULL, y = NULL, waveform = c("sine", "square", "triangle",
  "sawtooth"), ticks = NULL, tick_len = 0.05, pulse_len = 0,
  pulse_amp = 0.2, interpolation = c("spline", "linear", "constant"),
  noise_interval = c(0, 0), noise_amp = 0.5, duration = 5,
  amp_level = 1, stereo = TRUE, smp_rate = 44100, flim = c(440, 880),
  na_freq = 300, play = TRUE)
}
\arguments{
\item{x}{The x values. Can be used when y values are unevenly spaced. Default is -length(y)/2:length(y)/2}

\item{y}{The data values used to modulate the frequency.}

\item{waveform}{The waveform used for the sound. One of 'sine', 'square', 'triangle', 'sawtooth'. Default is 'sine'.}

\item{ticks}{The location of x-axis ticks, indicated with short bursts of a sawtooth wave (duration set by `tick_len`). The default is NULL (no ticks).}

\item{tick_len}{The duration of each tick sound.}

\item{pulse_len}{Length of individual pulses in seconds to mark the x-values. Default is 0.}

\item{pulse_amp}{Amplitude of pulses between 0 and 1. Default is 0.2.}

\item{interpolation}{The frequency interpolation method to connect the y-values. One of 'spline', 'linear', 'constant'. If 'constant', y[1] is played from x[1] to x[2], y[2] is played from x[2] to x[3], etc, and y[n] is played for the duration x[n] - x[n-1]. Default is 'spline'.}

\item{noise_interval}{White noise is overlayed whenever y is inside this interval (if noise_amp > 0) or outside this interval (if noise_amp < 0). For example, set to c(-Inf, 0) to indicate data in the negative range. Default is c(0,0) (no noise).}

\item{noise_amp}{Amplitude (between 0 and 1) of the noise used for noise_interval. Negative values (between 0 and -1) invert noise_interval. Default is 0.5.}

\item{duration}{Total duration of the generated sound in seconds. Default is 5.}

\item{amp_level}{Amplitude level between 0 and 1 to adjust the volume. Default is 1.}

\item{stereo}{If TRUE a left-to-right transition is simulated. Default is TRUE.}

\item{smp_rate}{The sampling rate of the wav file. Default is 44100 (CD quality)}

\item{flim}{The frequency range in Hz to which the data is mapped. Default is c(440, 880).}

\item{na_freq}{Frequency in Hz that is used for NA data. Default is 300.}

\item{play}{If TRUE, the sound is played. Default is TRUE.}
}
\value{
The synthesized sound saved as a tuneR::WaveMC object.
}
\description{
Synthesise a sound with frequency modulated by input data
}
\section{Licence}{

GPL (>=2)
}
\examples{
x = seq(-3, 3, .01)
y = dnorm(x)
sonify(y)

}
\author{
Stefan Siegert \email{s.siegert@exeter.ac.uk}
}
\seealso{
tuneR::play, tuneR::WaveMC
}

